Refactor: SideNavigationLayout pages to components  폴더 위치 변경#51
