+++ src/components/HistoryList.tsx
@@ -11,7 +11,7 @@ interface HistoryListProp {
 
 const HistoryList = ({ userHistoryData }: HistoryListProp) => {
   return (
-    <table className="w-full ">
+    <table className="w-full">
       <thead className="bg-blue-500 text-white text-xl rounded-t-lg sticky top-0">
         <tr>
           <th className="w-80 table-cell text-center p-5">위치</th>
